app/requirements.txt
pyserial
pynput
requests
beautifulsoup4
schedule
textblob
fastapi
uvicorn
python-dotenv  # For env vars later
praw